verl_distillation/requirements.txt
# requirements.txt records the full set of dependencies for development
accelerate
codetiming
click==8.0.4
datasets
dill
# flash-attn
hydra-core
liger-kernel
numpy<2.0.0
pandas
peft
pyarrow>=19.0.0
pybind11
pylatexenc
pre-commit
ray==2.49.0
tensordict>=0.8.0,<=0.9.1,!=0.9.0
torchdata
transformers
# vllm==0.8.4
opentelemetry-api>=1.26.0,<1.27.0
opentelemetry-sdk>=1.26.0,<1.27.0
opentelemetry-exporter-otlp-proto-grpc>=1.26.0,<1.27.0
opentelemetry-exporter-otlp-proto-http>=1.26.0,<1.27.0
wandb
packaging>=20.0
uvicorn
fastapi
latex2sympy2_extended
math_verify
sglang==0.5.2
